Convert search interceptor from string injection to real module
